Server: Change API. Some refactoring.
