+++ Server/Web/WebSocketService/Handlers/Authorization.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Security.Cryptography;
 using System.Text;
 
-namespace KeyBoardChat.Web.WebSocketService.Handler
+namespace KeyBoardChat.Web.WebSocketService.Handlers
 {
     public class AuthorizationHandler : WebSocketServiceHandler
     {
