+++ Server/Web/WebSocketService/Handlers/CreateRoom.cs
@@ -4,7 +4,7 @@ using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 
-namespace KeyBoardChat.Web.WebSocketService.Handler
+namespace KeyBoardChat.Web.WebSocketService.Handlers
 {
     public class CreateRoomHandler : WebSocketServiceHandler
     {
