+++ Server/Web/WebSocketService/Handlers/GetRooms.cs
@@ -4,7 +4,7 @@ using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 
-namespace KeyBoardChat.Web.WebSocketService.Handler
+namespace KeyBoardChat.Web.WebSocketService.Handlers
 {
     public class GetRoomsHandler : WebSocketServiceHandler
     {
