+++ Server/Web/WebSocketService/Handlers/GetUsers.cs
@@ -8,7 +8,7 @@ using System.IO;
 using System.Linq;
 using System.Text;
 
-namespace KeyBoardChat.Web.WebSocketService.Handler
+namespace KeyBoardChat.Web.WebSocketService.Handlers
 {
     public class GetUsersHandler : WebSocketServiceHandler
     {
