+++ Server/Web/WebSocketService/Handlers/Registration.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Security.Cryptography;
 using System.Text;
 
-namespace KeyBoardChat.Web.WebSocketService.Handler
+namespace KeyBoardChat.Web.WebSocketService.Handlers
 {
     public class RegistrationHandler : WebSocketServiceHandler
     {
